move data processing helpers to data_util
